+++ src/store/index.js
@@ -4,6 +4,7 @@ import headerStore from "@/store/modules/headerStore.js";
 import userStore from "@/store/modules/userStore.js";
 import modalStore from "@/store/modules/modalStore.js";
 import notificationStore from "@/store/modules/notificationStore.js";
+import navigationStore from "@/store/modules/navigationStore.js";
 
 export const store = createStore({
   modules: {
@@ -11,6 +12,7 @@ export const store = createStore({
     authStore,
     userStore,
     notificationStore,
-    modalStore
+    modalStore,
+    navigationStore
   },
 });
