+++ src/store/modules/navigationStore.js
@@ -1,0 +1,22 @@
+const state = {
+  currentTab: 'NotificationList'
+};
+
+const mutations = {
+  setCurrentTab(state, tabName) {
+    state.currentTab = tabName;
+  }
+};
+
+const actions = {
+  changeTab({commit}, tabName) {
+    commit('setCurrentTab', tabName);
+  }
+};
+
+export default {
+  namespaced: true,
+  state,
+  mutations,
+  actions
+};
